+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/ServiceBusReceiverClient.java
@@ -7,7 +7,8 @@ import com.azure.core.annotation.ServiceClient;
 import com.azure.core.util.IterableStream;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.servicebus.ServiceBusClientBuilder.ServiceBusReceiverClientBuilder;
-import com.azure.messaging.servicebus.implementation.ServiceBusReceiverTracer;
+import com.azure.messaging.servicebus.implementation.instrumentation.ServiceBusReceiverInstrumentation;
+import com.azure.messaging.servicebus.implementation.instrumentation.ServiceBusTracer;
 import com.azure.messaging.servicebus.models.AbandonOptions;
 import com.azure.messaging.servicebus.models.CompleteOptions;
 import com.azure.messaging.servicebus.models.DeadLetterOptions;
@@ -17,7 +18,6 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Sinks;
 
 import java.time.Duration;
-import java.time.Instant;
 import java.time.OffsetDateTime;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -66,7 +66,7 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
     /* To ensure synchronousMessageSubscriber is subscribed only once. */
     private final AtomicBoolean syncSubscribed = new AtomicBoolean(false);
 
-    private final ServiceBusReceiverTracer tracer;
+    private final ServiceBusTracer tracer;
     /**
      * Creates a synchronous receiver given its asynchronous counterpart.
      *
@@ -80,7 +80,7 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
         this.asyncClient = Objects.requireNonNull(asyncClient, "'asyncClient' cannot be null.");
         this.operationTimeout = Objects.requireNonNull(operationTimeout, "'operationTimeout' cannot be null.");
         this.isPrefetchDisabled = isPrefetchDisabled;
-        this.tracer = asyncClient.getTracer();
+        this.tracer = asyncClient.getInstrumentation().getTracer();
     }
 
     /**
@@ -396,11 +396,10 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
                 "'maxMessages' cannot be less than or equal to 0. maxMessages: " + maxMessages));
         }
 
-        final Instant startTime = tracer.isEnabled() ? Instant.now() : null;
         final Flux<ServiceBusReceivedMessage> messages = asyncClient.peekMessages(maxMessages, sessionId)
             .timeout(operationTimeout);
 
-        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.reportSyncReceiverSpan("ServiceBus.peekMessages", startTime, messages);
+        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.peekMessages", messages);
         // Subscribe to message flux so we can kick off this operation, but not to tracing - caller subscriber will take care of it.
         messages.subscribe();
 
@@ -448,11 +447,10 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
                 "'maxMessages' cannot be less than or equal to 0. maxMessages: " + maxMessages));
         }
 
-        final Instant startTime = tracer.isEnabled() ? Instant.now() : null;
         final Flux<ServiceBusReceivedMessage> messages = asyncClient.peekMessages(maxMessages, sequenceNumber,
             sessionId).timeout(operationTimeout);
 
-        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.reportSyncReceiverSpan("ServiceBus.peekMessages", startTime, messages);
+        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.peekMessages", messages);
         // Subscribe to message flux so we can kick off this operation, but not to tracing - caller subscriber will take care of it.
         messages.subscribe();
 
@@ -506,8 +504,6 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
                 new IllegalArgumentException("'maxWaitTime' cannot be zero or less. maxWaitTime: " + maxWaitTime));
         }
 
-        final Instant startTime = tracer.isEnabled() ? Instant.now() : null;
-
         // There are two subscribers to this emitter. One is the timeout between messages subscription in
         // SynchronousReceiverWork.start() and the other is the IterableStream(emitter.asFlux());
         // Since the subscriptions may happen at different times, we want to replay results to downstream subscribers.
@@ -516,7 +512,7 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
         queueWork(maxMessages, maxWaitTime, emitter);
 
         final Flux<ServiceBusReceivedMessage> messagesFlux = emitter.asFlux();
-        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.reportSyncReceiverSpan("ServiceBus.receiveMessages", startTime, messagesFlux);
+        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.receiveMessages", messagesFlux);
 
         // Subscribe to message flux so we can kick off this operation, but not to tracing - caller subscriber will take care of it.
         messagesFlux.subscribe();
@@ -589,11 +585,10 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
     IterableStream<ServiceBusReceivedMessage> receiveDeferredMessageBatch(Iterable<Long> sequenceNumbers,
         String sessionId) {
 
-        final Instant startTime = tracer.isEnabled() ? Instant.now() : null;
         final Flux<ServiceBusReceivedMessage> messages = asyncClient.receiveDeferredMessages(sequenceNumbers,
             sessionId).timeout(operationTimeout);
 
-        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.reportSyncReceiverSpan("ServiceBus.receiveDeferredMessageBatch", startTime, messages);
+        final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.receiveDeferredMessageBatch", messages);
         // Subscribe so we can kick off this operation.
         messages.subscribe();