+++ sdk/servicebus/azure-messaging-servicebus/src/test/java/com/azure/messaging/servicebus/ServiceBusSenderAsyncClientTest.java
@@ -15,6 +15,9 @@ import com.azure.core.amqp.implementation.ErrorContextProvider;
 import com.azure.core.amqp.implementation.MessageSerializer;
 import com.azure.core.amqp.models.CbsAuthorizationType;
 import com.azure.core.credential.TokenCredential;
+import com.azure.core.test.utils.metrics.TestCounter;
+import com.azure.core.test.utils.metrics.TestMeasurement;
+import com.azure.core.test.utils.metrics.TestMeter;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.ClientOptions;
 import com.azure.core.util.Context;
@@ -53,8 +56,11 @@ import java.nio.ByteBuffer;
 import java.time.Duration;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.Matcher;
@@ -63,12 +69,14 @@ import java.util.stream.IntStream;
 
 import static com.azure.core.util.tracing.Tracer.AZ_TRACING_NAMESPACE_KEY;
 import static com.azure.core.util.tracing.Tracer.DIAGNOSTIC_ID_KEY;
+import static com.azure.core.util.tracing.Tracer.HOST_NAME_KEY;
 import static com.azure.core.util.tracing.Tracer.PARENT_TRACE_CONTEXT_KEY;
 import static com.azure.core.util.tracing.Tracer.SPAN_BUILDER_KEY;
 import static com.azure.core.util.tracing.Tracer.SPAN_CONTEXT_KEY;
 import static com.azure.messaging.servicebus.ServiceBusSenderAsyncClient.MAX_MESSAGE_LENGTH_BYTES;
 import static com.azure.messaging.servicebus.implementation.ServiceBusConstants.AZ_TRACING_NAMESPACE_VALUE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyList;
@@ -95,7 +103,7 @@ class ServiceBusSenderAsyncClientTest {
     private static final BinaryData TEST_CONTENTS = BinaryData.fromString("My message for service bus queue!");
     private static final String TXN_ID_STRING = "1";
     private static final String CLIENT_IDENTIFIER = "my-client-identifier";
-
+    private static final ServiceBusSenderInstrumentation DEFAULT_INSTRUMENTATION = new ServiceBusSenderInstrumentation(null, null, NAMESPACE, ENTITY_NAME);
     @Mock
     private AmqpSendLink sendLink;
     @Mock
@@ -129,7 +137,6 @@ class ServiceBusSenderAsyncClientTest {
     private ArgumentCaptor<Iterable<Long>> sequenceNumberCaptor;
 
     private final MessageSerializer serializer = new ServiceBusMessageSerializer();
-    private static final ServiceBusSenderTracer DEFAULT_TRACER = new ServiceBusSenderTracer(null, NAMESPACE, ENTITY_NAME);
     private final AmqpRetryOptions retryOptions = new AmqpRetryOptions()
         .setDelay(Duration.ofMillis(500))
         .setMode(AmqpRetryMode.FIXED)
@@ -167,7 +174,7 @@ class ServiceBusSenderAsyncClientTest {
                 connectionOptions.getRetry()));
 
         sender = new ServiceBusSenderAsyncClient(ENTITY_NAME, MessagingEntityType.QUEUE, connectionProcessor,
-            retryOptions, DEFAULT_TRACER, serializer, onClientClose, null, CLIENT_IDENTIFIER);
+            retryOptions, DEFAULT_INSTRUMENTATION, serializer, onClientClose, null, CLIENT_IDENTIFIER);
 
         when(connection.getManagementNode(anyString(), any(MessagingEntityType.class)))
             .thenReturn(just(managementNode));
@@ -324,7 +331,7 @@ class ServiceBusSenderAsyncClientTest {
         final int count = 4;
         final byte[] contents = TEST_CONTENTS.toBytes();
         final ServiceBusMessageBatch batch = new ServiceBusMessageBatch(256 * 1024,
-            errorContextProvider, DEFAULT_TRACER, serializer, null, null);
+            errorContextProvider, DEFAULT_INSTRUMENTATION.getTracer(), serializer);
 
         IntStream.range(0, count).forEach(index -> {
             final ServiceBusMessage message = new ServiceBusMessage(BinaryData.fromBytes(contents));
@@ -363,7 +370,7 @@ class ServiceBusSenderAsyncClientTest {
         final int count = 4;
         final byte[] contents = TEST_CONTENTS.toBytes();
         final ServiceBusMessageBatch batch = new ServiceBusMessageBatch(256 * 1024,
-            errorContextProvider, DEFAULT_TRACER, serializer, null, null);
+            errorContextProvider, DEFAULT_INSTRUMENTATION.getTracer(), serializer);
 
         IntStream.range(0, count).forEach(index -> {
             final ServiceBusMessage message = new ServiceBusMessage(BinaryData.fromBytes(contents));
@@ -386,22 +393,19 @@ class ServiceBusSenderAsyncClientTest {
         messagesSent.forEach(message -> Assertions.assertEquals(Section.SectionType.Data, message.getBody().getType()));
     }
 
-    /**
-     * Verifies that sending multiple message will result in calling sender.send(MessageBatch).
-     */
     @Test
-    void sendMultipleMessagesTracerSpans() {
+    void sendMultipleMessagesTracesSpans() {
         // Arrange
         final int count = 4;
         final byte[] contents = TEST_CONTENTS.toBytes();
         final Tracer tracer1 = mock(Tracer.class);
         String traceparent = "traceparent";
-        ServiceBusSenderTracer sbTracer1 = new ServiceBusSenderTracer(tracer1, NAMESPACE, ENTITY_NAME);
+        ServiceBusSenderInstrumentation instrumentation = new ServiceBusSenderInstrumentation(tracer1, null, NAMESPACE, ENTITY_NAME);
 
         final ServiceBusMessageBatch batch = new ServiceBusMessageBatch(256 * 1024,
-            errorContextProvider, sbTracer1, serializer, null, null);
+            errorContextProvider, instrumentation.getTracer(), serializer);
         sender = new ServiceBusSenderAsyncClient(ENTITY_NAME, MessagingEntityType.QUEUE, connectionProcessor,
-            retryOptions, sbTracer1, serializer, onClientClose, null, CLIENT_IDENTIFIER);
+            retryOptions, instrumentation, serializer, onClientClose, null, CLIENT_IDENTIFIER);
 
         when(connection.createSendLink(eq(ENTITY_NAME), eq(ENTITY_NAME), eq(retryOptions), isNull(), eq(CLIENT_IDENTIFIER)))
             .thenReturn(Mono.just(sendLink));
@@ -447,6 +451,138 @@ class ServiceBusSenderAsyncClientTest {
         verify(tracer1, times(1))
             .start(eq("ServiceBus.send"), any(Context.class), eq(ProcessKind.SEND));
         verify(tracer1, times(5)).end(eq("success"), isNull(), any(Context.class));
+    }
+
+    @Test
+    void sendMessageReportsMetrics() {
+        // Arrange
+        TestMeter meter = new TestMeter();
+        ServiceBusSenderInstrumentation instrumentation = new ServiceBusSenderInstrumentation(null, meter, NAMESPACE, ENTITY_NAME);
+
+        sender = new ServiceBusSenderAsyncClient(ENTITY_NAME, MessagingEntityType.QUEUE, connectionProcessor,
+            retryOptions, instrumentation, serializer, onClientClose, null, CLIENT_IDENTIFIER);
+
+        when(connection.createSendLink(eq(ENTITY_NAME), eq(ENTITY_NAME), eq(retryOptions), isNull(), eq(CLIENT_IDENTIFIER)))
+            .thenReturn(Mono.just(sendLink));
+        when(sendLink.send(any(Message.class))).thenReturn(Mono.empty());
+
+        // Act
+        StepVerifier.create(sender.sendMessage(new ServiceBusMessage(TEST_CONTENTS))
+                .then(sender.sendMessage(new ServiceBusMessage(TEST_CONTENTS))))
+            .verifyComplete();
+
+        // Assert
+        TestCounter sentMessagesCounter = meter.getCounters().get("messaging.servicebus.messages.sent");
+        assertNotNull(sentMessagesCounter);
+        assertEquals(2, sentMessagesCounter.getMeasurements().size());
+
+        TestMeasurement<Long> measurement1 = sentMessagesCounter.getMeasurements().get(0);
+        TestMeasurement<Long> measurement2 = sentMessagesCounter.getMeasurements().get(1);
+        assertEquals(1, measurement1.getValue());
+        assertEquals(1, measurement2.getValue());
+
+        Map<String, Object> attributes1 = measurement1.getAttributes();
+        Map<String, Object> attributes2 = measurement2.getAttributes();
+        assertEquals(3, attributes1.size());
+        assertCommonMetricAttributes(attributes1, true);
+        assertEquals(3, attributes2.size());
+        assertCommonMetricAttributes(attributes2, true);
+    }
+
+    @Test
+    void sendMessageReportsMetricsAndTraces() {
+        // Arrange
+        TestMeter meter = new TestMeter();
+        Tracer tracer = mock(Tracer.class);
+        ServiceBusSenderInstrumentation instrumentation = new ServiceBusSenderInstrumentation(tracer, meter, NAMESPACE, ENTITY_NAME);
+
+        sender = new ServiceBusSenderAsyncClient(ENTITY_NAME, MessagingEntityType.QUEUE, connectionProcessor,
+            retryOptions, instrumentation, serializer, onClientClose, null, CLIENT_IDENTIFIER);
+
+        when(connection.createSendLink(eq(ENTITY_NAME), eq(ENTITY_NAME), eq(retryOptions), isNull(), eq(CLIENT_IDENTIFIER)))
+            .thenReturn(Mono.just(sendLink));
+        when(sendLink.send(any(Message.class))).thenReturn(Mono.empty());
+
+        Context span = new Context("marker", true);
+        when(tracer.start(eq("ServiceBus.send"), any(Context.class), eq(ProcessKind.SEND)))
+            .thenReturn(span);
+
+        when(tracer.extractContext(any(), any(Context.class))).thenReturn(Context.NONE);
+        when(tracer.start(eq("ServiceBus.message"), any(Context.class), any())).thenReturn(Context.NONE);
+        when(tracer.getSharedSpanBuilder(eq("ServiceBus.send"), any(Context.class))).thenReturn(Context.NONE);
+
+        // Act
+        StepVerifier.create(sender.sendMessage(new ServiceBusMessage(TEST_CONTENTS)))
+            .verifyComplete();
+
+        // Assert
+        TestCounter sentMessagesCounter = meter.getCounters().get("messaging.servicebus.messages.sent");
+        TestMeasurement<Long> measurement = sentMessagesCounter.getMeasurements().get(0);
+        assertEquals(1, measurement.getValue());
+
+        Map<String, Object> attributes = measurement.getAttributes();
+        assertEquals(3, attributes.size());
+        assertCommonMetricAttributes(attributes, true);
+        assertEquals(span, measurement.getContext());
+    }
+
+    @Test
+    void sendMessageBatchReportsMetrics() {
+        // Arrange
+        TestMeter meter = new TestMeter();
+        ServiceBusSenderInstrumentation instrumentation = new ServiceBusSenderInstrumentation(null, meter, NAMESPACE, ENTITY_NAME);
+
+        final ServiceBusMessageBatch batch = new ServiceBusMessageBatch(256 * 1024,
+            errorContextProvider, instrumentation.getTracer(), serializer);
+        batch.tryAddMessage(new ServiceBusMessage(TEST_CONTENTS));
+        batch.tryAddMessage(new ServiceBusMessage(TEST_CONTENTS));
+
+        sender = new ServiceBusSenderAsyncClient(ENTITY_NAME, MessagingEntityType.QUEUE, connectionProcessor,
+            retryOptions, instrumentation, serializer, onClientClose, null, CLIENT_IDENTIFIER);
+
+        when(connection.createSendLink(eq(ENTITY_NAME), eq(ENTITY_NAME), eq(retryOptions), isNull(), eq(CLIENT_IDENTIFIER)))
+            .thenReturn(Mono.just(sendLink));
+        when(sendLink.send(anyList())).thenReturn(Mono.empty());
+
+        // Act
+        StepVerifier.create(sender.sendMessages(batch))
+            .verifyComplete();
+
+        // Assert
+        TestCounter sentMessagesCounter = meter.getCounters().get("messaging.servicebus.messages.sent");
+        TestMeasurement<Long> measurement = sentMessagesCounter.getMeasurements().get(0);
+        assertEquals(2, measurement.getValue());
+
+        assertEquals(3,  measurement.getAttributes().size());
+        assertCommonMetricAttributes( measurement.getAttributes(), true);
+    }
+
+    @Test
+    void failedSendMessageReportsMetrics() {
+        // Arrange
+        TestMeter meter = new TestMeter();
+        ServiceBusSenderInstrumentation instrumentation = new ServiceBusSenderInstrumentation(null, meter, NAMESPACE, ENTITY_NAME);
+
+        sender = new ServiceBusSenderAsyncClient(ENTITY_NAME, MessagingEntityType.QUEUE, connectionProcessor,
+            retryOptions, instrumentation, serializer, onClientClose, null, CLIENT_IDENTIFIER);
+
+        when(connection.createSendLink(eq(ENTITY_NAME), eq(ENTITY_NAME), eq(retryOptions), isNull(), eq(CLIENT_IDENTIFIER)))
+            .thenReturn(Mono.just(sendLink));
+        when(sendLink.send(any(Message.class))).thenThrow(new RuntimeException("foo"));
+
+        // Act
+        StepVerifier.create(sender.sendMessage(new ServiceBusMessage(TEST_CONTENTS)))
+            .expectError()
+            .verify();
+
+        // Assert
+        TestCounter sentMessagesCounter = meter.getCounters().get("messaging.servicebus.messages.sent");
+        TestMeasurement<Long> measurement = sentMessagesCounter.getMeasurements().get(0);
+        assertEquals(1, measurement.getValue());
+
+        Map<String, Object> attributes = measurement.getAttributes();
+        assertEquals(3, attributes.size());
+        assertCommonMetricAttributes(attributes, false);
     }
 
     /**
@@ -766,5 +902,11 @@ class ServiceBusSenderAsyncClientTest {
 
         // Assert
         verify(onClientClose).run();
+    }
+
+    private void assertCommonMetricAttributes(Map<String, Object> attributes, boolean success) {
+        assertEquals(NAMESPACE, attributes.get("hostName"));
+        assertEquals(ENTITY_NAME, attributes.get("entityName"));
+        assertEquals(success ? "ok" : "error", attributes.get("status"));
     }
 }